Refactor regex builtin character-class identifiers (#10814)
Refactors the builtin regex class integer ids to common header for the compiler and executor.
The builtin regex character classes like `\s, \d, \W` have integer identifiers (bit values that can be combined) but were defined in separate source files. This PR refactors the declarations to the common header file `regcomp.h` to ensure the same value is used when parsing/compiling the instructions in `regcomp.cpp` and when evaluating the instructions in `regex.inl`.

This is just a cleanup of the code and does not effect behavior or performance.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Karthikeyan (https://github.com/karthikeyann)
  - Mike Wilson (https://github.com/hyperbolic2346)

URL: https://github.com/rapidsai/cudf/pull/10814