Refactor `lists::contains` (#11019)
This is just a not-very-simple refactor to `lists::contains`:
 * Remove some wrong lines in doxygen of `lists/contains.hpp`, and rewrite doxygen there a little bit.
 * Add more comments to the code.
 * Reduce the number of code paths of the template struct functors.
 * Rename some/many variables, and reorganize code to make it cleaner.

No new feature is added in this PR, just modifying the existing functions and moving things around.

This PR is extracted from the bigger PR for easier review. The original PR is https://github.com/rapidsai/cudf/pull/10548 for supporting nested type in `lists::contains`. As such, this blocks it.

Authors:
  - Nghia Truong (https://github.com/ttnghia)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Karthikeyan (https://github.com/karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/11019