Remove Index.replace API (#11131)
This PR is a follow-up to #11073. Its sole purpose is to remove the deprecated public `replace` API for Index objects. A stripped down, minimal version of this logic has been inlined in `DataFrame.rename`, which is the only place that needs that functionality. The scope of this PR is intentionally limited so that further changes can happen in parallel with changes associated with the removal of `Frame._index`, which should no longer be necessary after this PR is merged (aside from its potential usage in internal Cython code that can be freely worked around).

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)

URL: https://github.com/rapidsai/cudf/pull/11131