Remove usage of Frame in StringMethods (#11181)
Various methods of StringMethods create a Frame from a libcudf method and pass that up to `_return_or_inplace`, which handles this as a special case where a method operating on a single column returns multiple columns (e.g. splitting a string column into multiple columns containing substrings). In all such cases, there is no need to pass the index through. When the parent object is an index, the index can simply be discarded, while if the parent is a Series the parent's index is always used for the new object. This removes various unnecessary constructions of the Frame class.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cudf/pull/11181