Fix a memory aliasing/crash issue in scatter for lists. (#11254)
Fixes:  https://github.com/rapidsai/cudf/issues/11248

A transform iterator was calling `copy_bitmask()` to create a new `rmm::device_buffer` as a temporary.  The pointer from this was being stored in a `lists_column_view`, but the underlying temporary was being deleted immediately.  

The copy is unnecessary because the backing bitmask is coming from column views that have a safe lifetime for the duration of the operation.  So the fix is to just pass the original pointers along.

Authors:
  - https://github.com/nvdbaranec

Approvers:
  - Michael Wang (https://github.com/isVoid)
  - David Wendt (https://github.com/davidwendt)
  - Nghia Truong (https://github.com/ttnghia)
  - Robert (Bobby) Evans (https://github.com/revans2)

URL: https://github.com/rapidsai/cudf/pull/11254