Adding byte_array statistics (#11303)
This PR is needed for PR #11160 and another PR for writing list<int8>'s as byte arrays as well.

This PR will be the basis for the statistics written by #11160, but it should be noted that it will work due to an implementation detail. The union of the statistics types for strings and byte arrays is the same layout. This means the string statistics that are stored from the string column can be read as byte array statistics later when writing out the file and it will work simply because they can alias to each other.

Alternatives considered included changing the incoming table type in the case of strings being written as a byte array to be a byte array, which would only work because of the implementation detail that the columns are identical data layouts. Also considered was templating out multiple paths in [column_statistics.cuh:107](https://github.com/rapidsai/cudf/blob/d350bed12055cb4b9cc8bbd96e58670ca175dcc3/cpp/src/io/statistics/column_statistics.cuh#L107) that we could use to branch based on the data type.

The latter seems more the correct way, but comes with a performance and complexity issue. The method used in the PR was mainly chosen due to simplicity and the existing reliance on the data formats being identical throughout the code. It seems unlikely these will ever decouple.

Authors:
  - Mike Wilson (https://github.com/hyperbolic2346)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - MithunR (https://github.com/mythrocks)
  - https://github.com/nvdbaranec

URL: https://github.com/rapidsai/cudf/pull/11303