Fix atomic operations on NaN values (#11420)
Generic atomic operations are currently implemented using an `atomicCAS` in a loop that determines when the current thread's result is the one that was actually saved to the address. The final check is performed by directly comparing the values, which can lead to infinite loops when the value being set is a `NaN` because all comparisons involving `NaN`s return false. This PR fixes that issues by casting the data to an integral type and comparing those, bypassing `NaN` comparison.

This error was discovered in in hash-based aggregates `min` and `max`, and fixing it is a blocker for https://github.com/NVIDIA/spark-rapids/pull/5989.

Authors:
   - Nghia Truong (https://github.com/ttnghia)
   - Bradley Dice (https://github.com/bdice)

Approvers:
   - Bradley Dice (https://github.com/bdice)
   - Vyas Ramasubramani (https://github.com/vyasr)