Fix JNI for TableWithMeta to use schema_info instead of column_names (#11566)
After #11364 `TableWithMeta` no longer returns the column names for a table since it is looking at the old `column_names` field instead of the newer `schema_info` field in the table metadata.  This updates the JNI to use the `schema_info` to get the column names and adds a test for this API which was missing before.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Ryan Lee (https://github.com/rwlee)
  - Peixin (https://github.com/pxLi)

URL: https://github.com/rapidsai/cudf/pull/11566