Remove use of CUDA driver API calls from libcudf (#11370)
This PR adds the final bits that remove `libcuda.so` as a shared library dependency of `libcudf`. Currently, that's just one place that uses the driver API `cuContextGetCurrent()` to get the current CUDA context and maintain a per-context singleton. It was determined in the discussions within this PR that that could be replaced with a per-device singleton and a call to the runtime API `cudaGetDevice()` instead.

Authors:
  - Ashwin Srinath (https://github.com/shwina)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - David Wendt (https://github.com/davidwendt)
  - Jake Hemstad (https://github.com/jrhemstad)
  - Wonchan Lee (https://github.com/magnatelee)

URL: https://github.com/rapidsai/cudf/pull/11370